chore: extract styled-components from StakeFlowModal
